chore: implement contribution amount
