+++ app/src/store/cart.ts
@@ -23,14 +23,13 @@ import {
   BytesLike,
   Contract,
   ContractTransaction,
-  formatUnits,
   hexDataSlice,
   MaxUint256,
   isAddress,
   parseUnits,
   getAddress,
 } from 'src/utils/ethers';
-import { formatNumber } from 'src/utils/utils';
+import { assertSufficientBalance } from 'src/utils/utils';
 import useDataStore from 'src/store/data';
 import useWalletStore from 'src/store/wallet';
 
@@ -360,30 +359,4 @@ async function quoteExactInput(path: BytesLike, amountIn: BigNumberish): Promise
   const quoter = new Contract('0xb27308f9F90D607463bb33eA1BeBb41C27CE5AB6', abi, provider.value);
   const amountOut = await quoter.quoteExactInput(path, amountIn);
   return amountOut.mul(995).div(1000); // multiplying by 995/1000 is equivalent to 0.5% slippage
-}
-
-/**
- * @notice Throws an error if the user does not have sufficient balance for the specified token
- * @param tokenAddress Address of the token to check
- * @param requiredAmount Amount required as a raw number, e.g. 5e18 if the user needs 5 DAI
- * @returns True if the user has sufficient balance, or throws otherwise
- */
-async function assertSufficientBalance(tokenAddress: string, requiredAmount: BigNumberish): Promise<boolean> {
-  const { provider, userAddress } = useWalletStore();
-  if (!userAddress.value) return true; // exit early, don't want any errors thrown
-  const isEth = tokenAddress === WETH_ADDRESS;
-  tokenAddress = isEth ? ETH_ADDRESS : tokenAddress;
-  const abi = ['function balanceOf(address) view returns (uint256)'];
-  const token = new Contract(tokenAddress, abi, provider.value);
-  const balance = isEth ? await provider.value.getBalance(userAddress.value) : await token.balanceOf(userAddress.value);
-  if (balance.lt(requiredAmount)) {
-    const tokenInfo = SUPPORTED_TOKENS_MAPPING[tokenAddress];
-    const { symbol, decimals } = tokenInfo;
-    const balanceNeeds = formatNumber(formatUnits(requiredAmount, decimals), 4);
-    const balanceHas = formatNumber(formatUnits(balance, decimals), 4);
-    throw new Error(
-      `Insufficient ${symbol} balance: Current cart requires ${balanceNeeds} ${symbol}, but you only have ${balanceHas} ${symbol}`
-    );
-  }
-  return true;
 }
